Day 03
My first solution for part 2 (with conditionals) just had an extended
regex pattern matching all three commands, but not only was the regex
much less readable, picking the right capture groups was error-prone.
So I switched to this scanning approach instead, which might be less
efficient and more complex overall, but should be somewhat more clear.