Day 04
Didn't have a good time with this one. There's probably a far more
elegant way to deal with the wordsearch part, but diagonal traversal
of 2d arrays makes my head spin. Luckily the second part was so
different that I got to ignore the code I already wrote, and found a
relatively simple solution.