Day 12
My first solution for part 1 calculated area and perimeter together so
only needed the one bool array. Part 2 involved quiet a bit of
refactoring as a result to avoid duplication.

Tried out using a 1d array to hold the 2d grid for fun. It worked pretty
well in general, simplifying a lot of loops, but did add extra complexity
when checking if indices are in bounds or at the edge in some direction.